requirements.txt
pymongo
redis
psycopg2-binary
meo
